requirements.txt
fastapi
uvicorn
gradio
pandas
requests
openai
pytrends
plotly
python-dotenv